Derivatives wordt een class
